Further reorganized file structure to reflect different components of the rate limiting function and the portara directive
