+++ server/portara/portaraDirective.ts
@@ -1,0 +1,54 @@
+const { SchemaDirectiveVisitor }: any = require('graphql-tools');
+import {
+  defaultFieldResolver,
+  GraphQLField,
+  GraphQLObjectType,
+} from 'graphql';
+const asyncRedis = require('async-redis');
+const client = asyncRedis.createClient();
+import rateLimiter from './rateLimiter'
+
+export class portaraSchemaDirective extends SchemaDirectiveVisitor {
+
+  async generateErrorMessage(limit, per, name, ip) {
+    const timeLeft = await client.ttl(`${ip}_${name}`)
+    let error = `You have exceeded the request limit of ${limit} for the type(s) '${name}' . You have ${timeLeft} seconds left until the next request can be made.`;
+    return error;
+  }
+
+  visitFieldDefinition(field: GraphQLField<any, any>, details) {
+    const { limit, per } = this.args;
+    const { resolve = defaultFieldResolver } = field;
+    field.resolve = async (...originalArgs) => {
+      const [object, args, context, info] = originalArgs;
+      const error = await this.generateErrorMessage(limit, per, info.fieldName, context.req.ip)
+      const underLimit = await rateLimiter(limit, per, context.req.ip, info.fieldName);
+      if (underLimit) {
+        return resolve(...originalArgs);
+      } else {
+        const error = await this.generateErrorMessage(limit, per, info.fieldName, context.req.ip)
+        return new Error(error)
+      };
+    };
+  }
+
+  visitObject(type: GraphQLObjectType) {
+    const { limit, per } = this.args;
+    const fields = type.getFields();
+    Object.values(fields).forEach((field) => {
+      const { resolve = defaultFieldResolver } = field;
+      if (!field.astNode!.directives!.some((directive) => directive.name.value === 'portara')) {
+        field.resolve = async (...originalArgs) => {
+          const [object, args, context, info] = originalArgs;
+          const underLimit = await rateLimiter(limit, per, context.req.ip, type.toString());
+          if (underLimit) {
+            return resolve(...originalArgs);
+          } else {
+            const error = await this.generateErrorMessage(limit, per, type.toString(), context.req.ip)
+            return new Error(error)
+          };
+        };
+      }
+    });
+  }
+}
